Refactor buildFilter to toFilterParams
